Moved Misfit to own module. Import * in __init__
